refactor: reorganize amaru-plutus for easier usage
