+++ crates/amaru-plutus/src/lib.rs
@@ -17,9 +17,7 @@ pub mod script_context;
 pub mod to_plutus_data;
 
 #[doc(hidden)]
-/// **Unstable internal APIs** - no guarantees, may change at any time.
-/// Only use if you're maintaining code in this workspace.
-pub mod unstable;
+pub mod context;
 
 use to_plutus_data::*;
