+++ crates/amaru-plutus/src/script_context/mod.rs
@@ -13,28 +13,17 @@
 // limitations under the License.
 
 use amaru_kernel::{
-    AddrKeyhash, Address, AddressError, AlonzoValue, AssetName, Bytes, CborWrap, Certificate,
-    ComputeHash, DatumHash, EraHistory, Hash, KeepRaw, KeyValuePairs, Lovelace, MemoizedDatum,
-    MemoizedScript, MemoizedTransactionOutput, MintedDatumOption, MintedScriptRef,
-    MintedTransactionBody, MintedTransactionOutput, MintedWitnessSet, NativeScript, Network,
-    NonEmptyKeyValuePairs, NonEmptySet, PlutusData, PlutusScript, PolicyId, Proposal,
-    ProposalIdAdapter, PseudoScript, Redeemer, RewardAccount, ScriptPurpose as RedeemerTag, Slot,
-    StakeCredential, StakePayload, TransactionId, TransactionInput, TransactionInputAdapter, Vote,
-    Voter, VotingProcedures, network::NetworkName, normalize_redeemers,
-    protocol_parameters::GlobalParameters,
+    Certificate, KeyValuePairs, MemoizedTransactionOutput, PlutusData, Redeemer, TransactionInput,
+    Voter,
 };
-use amaru_slot_arithmetic::{EraHistoryError, TimeMs};
-use itertools::Itertools;
-use std::{
-    cmp::Ordering,
-    collections::{BTreeMap, BTreeSet},
-    ops::Deref,
-};
-use thiserror::Error;
+use std::{collections::BTreeMap, ops::Deref};
 
 pub mod v1;
 pub mod v2;
 pub mod v3;
+
+pub mod tx_info;
+pub use tx_info::*;
 
 pub trait IsPrePlutusVersion3 {}
 impl IsPrePlutusVersion3 for PlutusVersion<1> {}
@@ -138,265 +127,6 @@ impl<'a> ScriptContext<'a> {
     }
 }
 
-/// An opaque type that represents the `TxInfo` field used in a [`ScriptContext`].
-///
-/// `TxInfo` is an in-memory representation of a Cardano transaction used in Plutus scripts.
-///
-/// Notably, it is not an exact mapping of the transaction on the ledger.
-/// For example, bootstrap addresses are skipped in the inputs, reference inputs, and outputs.
-pub struct TxInfo {
-    pub(crate) inputs: Vec<OutputRef>,
-    pub(crate) reference_inputs: Vec<OutputRef>,
-    pub(crate) outputs: Vec<TransactionOutput>,
-    pub(crate) fee: Lovelace,
-    pub(crate) mint: Mint,
-    pub(crate) certificates: Vec<Certificate>,
-    pub(crate) withdrawals: Withdrawals,
-    pub(crate) valid_range: TimeRange,
-    pub(crate) signatories: RequiredSigners,
-    pub(crate) redeemers: Redeemers<ScriptPurpose>,
-    pub(crate) data: Datums,
-    pub(crate) id: TransactionId,
-    pub(crate) votes: Votes,
-    pub(crate) proposal_procedures: Vec<Proposal>,
-    pub(crate) current_treasury_amount: Option<Lovelace>,
-    pub(crate) treasury_donation: Option<Lovelace>,
-}
-
-#[derive(Debug, Error)]
-/// Represents possible errors that can occur during [`TxInfo` construction](TxInfo::new).
-///
-/// An occurance of this error should suggest a user error of one of two types:
-/// - A poorly constructed transaction that should fail phase-one validation
-/// - Incorrect chain state such as an incomplete UTxO slice, wrong network, or wrong slot value
-pub enum TxInfoTranslationError {
-    /// Some input was not in the provided [`Utxos`]
-    #[error("missing input: {0}")]
-    MissingInput(TransactionInputAdapter),
-    /// Some output is poorly constructed
-    #[error("invalid output: {0}")]
-    InvalidOutput(#[from] TransactionOutputError),
-    /// Some withdrawal is poorly constructed
-    #[error("invalid withdrawal: {0}")]
-    InvalidWithdrawal(#[from] WithdrawalError),
-    /// The validity interval cannot be converted to posix time
-    #[error("invalid validity interval: {0}")]
-    InvalidValidityInterval(#[from] EraHistoryError),
-    /// Some redeemer is poorly constructed
-    #[error("invalid redeemer at index {0}")]
-    InvalidRedeemer(usize),
-}
-
-impl TxInfo {
-    /// Construct a new `TxInfo` from a transaction and some additional context.
-    ///
-    /// This is a fallible operation which fails when some state can't be represented by `TxInfo`
-    /// See [TxInfoTranslationError] for more.
-    ///
-    /// It's important to note that a successful construction of a `TxInfo` does not mean it can be serialized for all Plutus versions.
-    /// For example, in Plutus V1, inputs that are locked by a bootstrap address are ignored, where as in V2 and V3 they are forbidden, resulting in an error.
-    ///
-    ///
-    /// Version-specific errors will arise during serialization.
-    pub fn new(
-        tx: MintedTransactionBody<'_>,
-        witness_set: MintedWitnessSet<'_>,
-        tx_id: &Hash<32>,
-        utxos: &Utxos,
-        slot: &Slot,
-        network: NetworkName,
-        era_history: &EraHistory,
-    ) -> Result<Self, TxInfoTranslationError> {
-        let inputs = Self::translate_inputs(tx.inputs.to_vec(), utxos)?;
-        let reference_inputs = tx
-            .reference_inputs
-            .map(|ref_inputs| Self::translate_inputs(ref_inputs.to_vec(), utxos))
-            .transpose()?
-            .unwrap_or_default();
-
-        let outputs = tx
-            .outputs
-            .into_iter()
-            .map(TransactionOutput::try_from)
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let mint = tx.mint.map(Mint::from).unwrap_or_default();
-
-        let certificates: Vec<Certificate> = tx
-            .certificates
-            .map(|certificates| certificates.to_vec())
-            .unwrap_or_default();
-
-        let withdrawals = tx
-            .withdrawals
-            .as_ref()
-            .map(Withdrawals::try_from)
-            .transpose()?
-            .unwrap_or_default();
-
-        let valid_range = TimeRange::new(
-            tx.validity_interval_start.map(Slot::from),
-            tx.ttl.map(Slot::from),
-            slot,
-            era_history,
-            network,
-        )?;
-
-        let signatories = tx
-            .required_signers
-            .as_ref()
-            .map(RequiredSigners::from)
-            .unwrap_or_default();
-
-        let proposal_procedures: Vec<_> = tx
-            .proposal_procedures
-            .map(|proposals| proposals.to_vec())
-            .unwrap_or_default();
-
-        let votes = tx.voting_procedures.map(Votes::from).unwrap_or_default();
-
-        let redeemers = Redeemers(
-            witness_set
-                .redeemer
-                .map(|redeemers| {
-                    normalize_redeemers(redeemers.unwrap())
-                        .into_iter()
-                        .enumerate()
-                        .map(|(ix, redeemer)| {
-                            let purpose = ScriptPurpose::builder(
-                                &redeemer,
-                                &inputs[..],
-                                &mint,
-                                &withdrawals,
-                                &certificates,
-                                &proposal_procedures,
-                                &votes,
-                            )
-                            .ok_or(TxInfoTranslationError::InvalidRedeemer(ix))?;
-
-                            Ok((purpose, redeemer))
-                        })
-                        .collect::<Result<Vec<(ScriptPurpose, Redeemer)>, TxInfoTranslationError>>()
-                })
-                .transpose()?
-                .unwrap_or_default(),
-        );
-
-        let data = witness_set
-            .plutus_data
-            .map(Datums::from)
-            .unwrap_or_default();
-
-        Ok(Self {
-            inputs,
-            reference_inputs,
-            outputs,
-            fee: tx.fee,
-            mint,
-            certificates,
-            withdrawals,
-            valid_range,
-            signatories,
-            redeemers,
-            data,
-            id: *tx_id,
-            votes,
-            proposal_procedures,
-            current_treasury_amount: tx.treasury_value,
-            treasury_donation: tx.donation.map(|donation| donation.into()),
-        })
-    }
-
-    fn translate_inputs(
-        inputs: Vec<TransactionInput>,
-        utxos: &Utxos,
-    ) -> Result<Vec<OutputRef>, TxInfoTranslationError> {
-        inputs
-            .into_iter()
-            .sorted()
-            .map(|input| {
-                utxos
-                    .resolve_input(&input)
-                    .ok_or(TxInfoTranslationError::MissingInput(input.clone().into()))
-            })
-            .collect::<Result<Vec<_>, _>>()
-    }
-}
-
-#[doc(hidden)]
-pub type ScriptPurpose = ScriptInfo<()>;
-
-#[doc(hidden)]
-#[derive(Clone)]
-pub enum ScriptInfo<T: Clone> {
-    Minting(PolicyId),
-    Spending(TransactionInput, T),
-    Rewarding(StakeCredential),
-    Certifying(usize, Certificate),
-    Voting(Voter),
-    Proposing(usize, Proposal),
-}
-
-impl ScriptPurpose {
-    pub fn builder(
-        redeemer: &Redeemer,
-        inputs: &[OutputRef],
-        mint: &Mint,
-        withdrawals: &Withdrawals,
-        certs: &[Certificate],
-        proposal_procedures: &[Proposal],
-        votes: &Votes,
-    ) -> Option<Self> {
-        let index = redeemer.index as usize;
-        match redeemer.tag {
-            RedeemerTag::Spend => inputs
-                .get(index)
-                .map(|output_ref| ScriptPurpose::Spending(output_ref.input.clone(), ())),
-            RedeemerTag::Mint => mint
-                .0
-                .keys()
-                .nth(index)
-                .copied()
-                .map(ScriptPurpose::Minting),
-            RedeemerTag::Reward => withdrawals.0.keys().nth(index).map(|stake| {
-                ScriptPurpose::Rewarding(match stake.0.payload() {
-                    amaru_kernel::StakePayload::Stake(hash) => StakeCredential::AddrKeyhash(*hash),
-                    amaru_kernel::StakePayload::Script(hash) => StakeCredential::ScriptHash(*hash),
-                })
-            }),
-            RedeemerTag::Cert => certs
-                .get(index)
-                .map(|cert| ScriptPurpose::Certifying(index, cert.clone())),
-            RedeemerTag::Vote => votes
-                .0
-                .keys()
-                .nth(index)
-                .map(|voter| ScriptPurpose::Voting(voter.clone())),
-            RedeemerTag::Propose => proposal_procedures
-                .get(index)
-                .map(|proposal| ScriptPurpose::Proposing(index, proposal.clone())),
-        }
-    }
-
-    pub fn to_script_info(self, data: Option<PlutusData>) -> ScriptInfo<Option<PlutusData>> {
-        match self {
-            ScriptInfo::Spending(input, _) => ScriptInfo::Spending(input, data),
-            ScriptInfo::Minting(policy) => ScriptInfo::Minting(policy),
-            ScriptInfo::Rewarding(stake_credential) => ScriptInfo::Rewarding(stake_credential),
-            ScriptInfo::Certifying(ix, certificate) => ScriptInfo::Certifying(ix, certificate),
-            ScriptInfo::Voting(voter) => ScriptInfo::Voting(voter),
-            ScriptInfo::Proposing(ix, proposal) => ScriptInfo::Proposing(ix, proposal),
-        }
-    }
-}
-
-/// A resolved input which includes the output it references.
-#[doc(hidden)]
-pub struct OutputRef {
-    pub input: TransactionInput,
-    pub output: TransactionOutput,
-}
-
 /// A subset of the UTxO set.
 ///
 /// Maps from a `TransactionInput` to a `MemoizedTransactionOutput`
@@ -426,499 +156,6 @@ impl Deref for Utxos {
 impl From<BTreeMap<TransactionInput, MemoizedTransactionOutput>> for Utxos {
     fn from(value: BTreeMap<TransactionInput, MemoizedTransactionOutput>) -> Self {
         Self(value)
-    }
-}
-
-/// An interval of time using POSIX time
-///
-///
-/// Time is a difficult, and heavily documented, challenge on Cardano.
-/// To maintain deterministic transaction validation,
-/// Cardano uses a validity interval which makes a transaction only valid from slot X to slot Y.
-///
-/// By default, the validity interval is unbounded, meaning the transcation could always be valid.
-///
-/// One wrinkle that this causes is that while Ouroboros uses slots to handle time, Plutus uses POSIX time.
-/// See [`TimeRange::new`] for more information on converting from a validity interval from Ouroboros to a Plutus TimeRange.
-#[doc(hidden)]
-pub struct TimeRange {
-    pub(crate) lower_bound: Option<TimeMs>,
-    pub(crate) upper_bound: Option<TimeMs>,
-}
-
-impl TimeRange {
-    /// Construct a new [`TimeRange`] given a slot interval
-    ///
-    /// Slot lengths can change based at hard forks, so it is not safe to count slots.
-    /// Conversion is handled by the provided `era_history`, which depends on the correct `network` (to determine the `GlobalParamters`)
-    ///
-    /// There are a few cases that would lead to an `EraHistoryError`:
-    /// - `EraHistoryError::PastTimeHorizon`:
-    ///   If a bound is too far in the future, we cannot be sure that a hardfork will occur that will change timings.
-    ///   The time horizon is the "stability window", which can take up to 3k/f
-    /// - `EraHistory::InvalidEraHistory`:
-    ///   One of the bounds cannot be found in any era in the `EraHistory`, so we do not know the slot length
-    ///   and thus cannot convert to POSIX time. This is typically going to be the result of a user error (incorrect era history)
-    pub fn new(
-        valid_from_slot: Option<Slot>,
-        valid_to_slot: Option<Slot>,
-        tip: &Slot,
-        era_history: &EraHistory,
-        network: NetworkName,
-    ) -> Result<Self, EraHistoryError> {
-        let parameters: &GlobalParameters = network.into();
-        let lower_bound = valid_from_slot
-            .map(|slot| era_history.slot_to_posix_time(slot, *tip, parameters.system_start.into()))
-            .transpose()?;
-        let upper_bound = valid_to_slot
-            .map(|slot| era_history.slot_to_posix_time(slot, *tip, parameters.system_start.into()))
-            .transpose()?;
-
-        Ok(Self {
-            lower_bound,
-            upper_bound,
-        })
-    }
-}
-
-/// An identifier for a currency in a [`Value`]
-///
-/// See [`Value`] for more
-#[doc(hidden)]
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum CurrencySymbol {
-    Lovelace,
-    Native(Hash<28>),
-}
-
-impl From<Hash<28>> for CurrencySymbol {
-    fn from(value: Hash<28>) -> Self {
-        Self::Native(value)
-    }
-}
-
-/// A representation of `Value` used in Plutus
-///
-/// The ledger's `Value` contains both a `Coin` and, optionally, a `Multiasset`.
-/// In Plutus, this is simply a single map, with an empty bytestring representing lovelace
-#[doc(hidden)]
-#[derive(Clone)]
-pub struct Value(pub BTreeMap<CurrencySymbol, BTreeMap<AssetName, u64>>);
-
-impl From<&amaru_kernel::Value> for Value {
-    fn from(value: &amaru_kernel::Value) -> Self {
-        let assets = match value {
-            amaru_kernel::Value::Coin(coin) => BTreeMap::from([(
-                CurrencySymbol::Lovelace,
-                BTreeMap::from([(Bytes::from(vec![]), *coin)]),
-            )]),
-            amaru_kernel::Value::Multiasset(coin, multiasset) => {
-                let mut map = BTreeMap::new();
-                map.insert(
-                    CurrencySymbol::Lovelace,
-                    BTreeMap::from([(Bytes::from(vec![]), *coin)]),
-                );
-                multiasset.iter().for_each(|(policy_id, asset_bundle)| {
-                    map.insert(
-                        CurrencySymbol::Native(*policy_id),
-                        asset_bundle
-                            .iter()
-                            .map(|(asset_name, amount)| (asset_name.clone(), amount.into()))
-                            .collect(),
-                    );
-                });
-
-                map
-            }
-        };
-
-        Self(assets)
-    }
-}
-
-impl From<amaru_kernel::Value> for Value {
-    fn from(value: amaru_kernel::Value) -> Self {
-        let assets = match value {
-            amaru_kernel::Value::Coin(coin) => BTreeMap::from([(
-                CurrencySymbol::Lovelace,
-                BTreeMap::from([(Bytes::from(vec![]), coin)]),
-            )]),
-            amaru_kernel::Value::Multiasset(coin, multiasset) => {
-                let mut map = BTreeMap::new();
-                map.insert(
-                    CurrencySymbol::Lovelace,
-                    BTreeMap::from([(Bytes::from(vec![]), coin)]),
-                );
-                multiasset
-                    .into_iter()
-                    .for_each(|(policy_id, asset_bundle)| {
-                        map.insert(
-                            CurrencySymbol::Native(policy_id),
-                            asset_bundle
-                                .into_iter()
-                                .map(|(asset_name, amount)| (asset_name, amount.into()))
-                                .collect(),
-                        );
-                    });
-
-                map
-            }
-        };
-
-        Self(assets)
-    }
-}
-
-impl From<Lovelace> for Value {
-    fn from(coin: Lovelace) -> Self {
-        Self(BTreeMap::from([(
-            CurrencySymbol::Lovelace,
-            BTreeMap::from([(Bytes::from(vec![]), coin)]),
-        )]))
-    }
-}
-
-#[doc(hidden)]
-#[derive(Debug, Error)]
-pub enum AlonzoValueError {
-    #[error("invalid quantity: {0}")]
-    InvalidQuantity(u64),
-}
-impl TryFrom<&AlonzoValue> for Value {
-    type Error = AlonzoValueError;
-
-    fn try_from(value: &AlonzoValue) -> Result<Self, Self::Error> {
-        let from_tokens = |tokens: &KeyValuePairs<AssetName, Lovelace>| {
-            (*tokens)
-                .iter()
-                .map(|(asset_name, quantity)| {
-                    if *quantity > 0 {
-                        Ok((asset_name.clone(), *quantity))
-                    } else {
-                        Err(AlonzoValueError::InvalidQuantity(*quantity))
-                    }
-                })
-                .collect::<Result<BTreeMap<_, _>, Self::Error>>()
-        };
-
-        match value {
-            AlonzoValue::Coin(coin) => Ok((*coin).into()),
-            AlonzoValue::Multiasset(coin, multiasset) if multiasset.is_empty() => {
-                Ok((*coin).into())
-            }
-            AlonzoValue::Multiasset(coin, multiasset) => {
-                let mut map = BTreeMap::new();
-                map.insert(
-                    CurrencySymbol::Lovelace,
-                    BTreeMap::from([(Bytes::from(vec![]), *coin)]),
-                );
-
-                for (policy_id, tokens) in multiasset.deref() {
-                    map.insert(CurrencySymbol::Native(*policy_id), from_tokens(tokens)?);
-                }
-
-                Ok(Self(map))
-            }
-        }
-    }
-}
-
-impl Value {
-    pub fn ada(&self) -> Option<u64> {
-        self.0
-            .get(&CurrencySymbol::Lovelace)
-            .and_then(|asset_bundle| {
-                asset_bundle.iter().find_map(|(name, amount)| {
-                    if name.is_empty() && amount != &0 {
-                        Some(*amount)
-                    } else {
-                        None
-                    }
-                })
-            })
-    }
-}
-
-#[doc(hidden)]
-#[derive(Clone)]
-pub enum Script {
-    Native(NativeScript),
-    PlutusV1(PlutusScript<1>),
-    PlutusV2(PlutusScript<2>),
-    PlutusV3(PlutusScript<3>),
-}
-
-impl From<CborWrap<MintedScriptRef<'_>>> for Script {
-    fn from(value: CborWrap<MintedScriptRef<'_>>) -> Self {
-        match value.0 {
-            PseudoScript::NativeScript(script) => Script::Native(script.unwrap()),
-            PseudoScript::PlutusV1Script(script) => Script::PlutusV1(script),
-            PseudoScript::PlutusV2Script(script) => Script::PlutusV2(script),
-            PseudoScript::PlutusV3Script(script) => Script::PlutusV3(script),
-        }
-    }
-}
-
-impl From<MemoizedScript> for Script {
-    fn from(value: MemoizedScript) -> Self {
-        match value {
-            PseudoScript::NativeScript(script) => Script::Native(script.as_ref().clone()),
-            PseudoScript::PlutusV1Script(script) => Script::PlutusV1(script),
-            PseudoScript::PlutusV2Script(script) => Script::PlutusV2(script),
-            PseudoScript::PlutusV3Script(script) => Script::PlutusV3(script),
-        }
-    }
-}
-
-#[doc(hidden)]
-#[derive(Clone)]
-pub enum DatumOption {
-    None,
-    Hash(DatumHash),
-    Inline(PlutusData),
-}
-
-impl From<MemoizedDatum> for DatumOption {
-    fn from(value: MemoizedDatum) -> Self {
-        match value {
-            MemoizedDatum::None => Self::None,
-            MemoizedDatum::Hash(hash) => Self::Hash(hash),
-            MemoizedDatum::Inline(data) => Self::Inline(data.as_ref().clone()),
-        }
-    }
-}
-
-impl From<Option<MintedDatumOption<'_>>> for DatumOption {
-    fn from(value: Option<MintedDatumOption<'_>>) -> Self {
-        match value {
-            None => Self::None,
-            Some(MintedDatumOption::Hash(hash)) => Self::Hash(hash),
-            Some(MintedDatumOption::Data(data)) => Self::Inline(data.0.unwrap()),
-        }
-    }
-}
-
-impl From<Option<&Hash<32>>> for DatumOption {
-    fn from(value: Option<&Hash<32>>) -> Self {
-        match value {
-            Some(hash) => Self::Hash(*hash),
-            None => Self::None,
-        }
-    }
-}
-
-#[doc(hidden)]
-#[derive(Clone)]
-pub struct TransactionOutput {
-    pub is_legacy: bool,
-    pub address: Address,
-    pub value: Value,
-    pub datum: DatumOption,
-    pub script: Option<Script>,
-}
-
-impl From<MemoizedTransactionOutput> for TransactionOutput {
-    fn from(output: MemoizedTransactionOutput) -> Self {
-        Self {
-            is_legacy: output.is_legacy,
-            address: output.address,
-            value: output.value.into(),
-            datum: output.datum.into(),
-            script: output.script.map(Script::from),
-        }
-    }
-}
-
-#[doc(hidden)]
-#[derive(Debug, Error)]
-pub enum TransactionOutputError {
-    #[error("invalid address: {0}")]
-    InvalidAddress(#[from] AddressError),
-    #[error("invalid value: {0}")]
-    InvalidValue(#[from] AlonzoValueError),
-}
-
-impl TryFrom<MintedTransactionOutput<'_>> for TransactionOutput {
-    type Error = TransactionOutputError;
-
-    fn try_from(output: MintedTransactionOutput<'_>) -> Result<TransactionOutput, Self::Error> {
-        match output {
-            MintedTransactionOutput::Legacy(output) => Ok(TransactionOutput {
-                is_legacy: true,
-                address: Address::from_bytes(&output.address)?,
-                value: (&output.amount).try_into()?,
-                datum: output.datum_hash.as_ref().into(),
-                script: None,
-            }),
-            MintedTransactionOutput::PostAlonzo(output) => Ok(TransactionOutput {
-                is_legacy: false,
-                address: Address::from_bytes(&output.address)?,
-                value: (&output.value).into(),
-                script: output.script_ref.map(Script::from),
-                datum: output.datum_option.into(),
-            }),
-        }
-    }
-}
-
-#[doc(hidden)]
-#[derive(Debug, Default)]
-pub struct Mint(pub BTreeMap<Hash<28>, BTreeMap<AssetName, i64>>);
-
-impl From<amaru_kernel::Mint> for Mint {
-    fn from(value: amaru_kernel::Mint) -> Self {
-        let mints = value
-            .into_iter()
-            .map(|(policy, multiasset)| {
-                (
-                    policy,
-                    multiasset
-                        .into_iter()
-                        .map(|(asset_name, amount)| (asset_name, amount.into()))
-                        .collect(),
-                )
-            })
-            .collect();
-
-        Self(mints)
-    }
-}
-
-#[doc(hidden)]
-#[derive(Default)]
-pub struct RequiredSigners(pub BTreeSet<AddrKeyhash>);
-
-impl<'a> From<&'a amaru_kernel::RequiredSigners> for RequiredSigners {
-    fn from(value: &'a amaru_kernel::RequiredSigners) -> Self {
-        Self(value.iter().copied().collect())
-    }
-}
-
-#[doc(hidden)]
-#[derive(Default)]
-pub struct Withdrawals(pub BTreeMap<StakeAddress, Lovelace>);
-
-#[doc(hidden)]
-#[derive(Clone, Debug)]
-pub struct StakeAddress(amaru_kernel::StakeAddress);
-
-impl From<StakeAddress> for amaru_kernel::StakeAddress {
-    fn from(value: StakeAddress) -> Self {
-        value.0
-    }
-}
-
-// Reference for this ordering is
-// https://github.com/aiken-lang/aiken/blob/a8c032935dbaf4a1140e9d8be5c270acd32c9e8c/crates/uplc/src/tx/script_context.rs#L1112
-impl Ord for StakeAddress {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        fn network_tag(network: Network) -> u8 {
-            match network {
-                Network::Testnet => 0,
-                Network::Mainnet => 1,
-                Network::Other(tag) => tag,
-            }
-        }
-
-        if self.0.network() != other.0.network() {
-            return network_tag(self.0.network()).cmp(&network_tag(other.0.network()));
-        }
-
-        match (self.0.payload(), other.0.payload()) {
-            (StakePayload::Script(..), StakePayload::Stake(..)) => Ordering::Less,
-            (StakePayload::Stake(..), StakePayload::Script(..)) => Ordering::Greater,
-            (StakePayload::Script(hash_a), StakePayload::Script(hash_b)) => hash_a.cmp(hash_b),
-            (StakePayload::Stake(hash_a), StakePayload::Stake(hash_b)) => hash_a.cmp(hash_b),
-        }
-    }
-}
-
-impl PartialOrd for StakeAddress {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for StakeAddress {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other) == Ordering::Equal
-    }
-}
-
-impl Eq for StakeAddress {}
-
-#[doc(hidden)]
-#[derive(Debug, Error)]
-pub enum WithdrawalError {
-    #[error("invalid reward account: {0}")]
-    InvalidRewardAccount(#[from] AddressError),
-    #[error("invalid address type: {0}")]
-    InvalidAddressType(Address),
-}
-
-impl TryFrom<&NonEmptyKeyValuePairs<RewardAccount, Lovelace>> for Withdrawals {
-    type Error = WithdrawalError;
-
-    fn try_from(
-        value: &NonEmptyKeyValuePairs<RewardAccount, Lovelace>,
-    ) -> Result<Self, Self::Error> {
-        let withdrawals = value
-            .iter()
-            .map(|(reward_account, coin)| {
-                let address = Address::from_bytes(reward_account)?;
-
-                if let Address::Stake(reward_account) = address {
-                    Ok((StakeAddress(reward_account), *coin))
-                } else {
-                    Err(WithdrawalError::InvalidAddressType(address))
-                }
-            })
-            .collect::<Result<BTreeMap<_, _>, WithdrawalError>>()?;
-
-        Ok(Self(withdrawals))
-    }
-}
-
-#[doc(hidden)]
-#[derive(Default)]
-pub struct Datums(pub BTreeMap<DatumHash, PlutusData>);
-
-impl From<NonEmptySet<KeepRaw<'_, PlutusData>>> for Datums {
-    fn from(plutus_data: NonEmptySet<KeepRaw<'_, PlutusData>>) -> Self {
-        Self(
-            plutus_data
-                .to_vec()
-                .into_iter()
-                .map(|data| (data.compute_hash(), data.unwrap()))
-                .collect(),
-        )
-    }
-}
-
-#[doc(hidden)]
-// FIXME: This should probably be a BTreeMap
-pub struct Redeemers<T>(pub Vec<(T, Redeemer)>);
-
-#[doc(hidden)]
-#[derive(Default)]
-pub struct Votes(pub BTreeMap<Voter, BTreeMap<ProposalIdAdapter, Vote>>);
-
-impl From<VotingProcedures> for Votes {
-    fn from(voting_procedures: VotingProcedures) -> Self {
-        Self(
-            voting_procedures
-                .into_iter()
-                .map(|(voter, votes)| {
-                    (
-                        voter,
-                        votes
-                            .into_iter()
-                            .map(|(proposal, procedure)| (proposal.into(), procedure.vote))
-                            .collect(),
-                    )
-                })
-                .collect(),
-        )
     }
 }
 
@@ -1140,7 +377,8 @@ pub mod test_vectors {
 
 #[cfg(test)]
 mod tests {
-    use amaru_kernel::{ShelleyAddress, ShelleyDelegationPart};
+
+    use amaru_kernel::{Bytes, Network, ShelleyAddress, ShelleyDelegationPart, StakePayload};
     use proptest::{
         prelude::{Just, Strategy, any, prop},
         prop_assert, prop_oneof, proptest,
